Change encoding name cases, following whatwg
'labels.py' is outdated. So I recreated it.
But in the meantime, whatwg introduced 'replacement' codec.
It should be dealt with a different commit,
so for this commit, I removed the new codec name and labels.
('mklabels.py')

See
Fix #32: use the encoding name exposed by DOM and HTML <form>
https://github.com/whatwg/encoding/commit/18bf4fa6f005a899696800a8f6b0d20f3a4c1c0e
Make the name of ISO-8859-8-I all uppercase in encodings.json
https://github.com/whatwg/encoding/commit/9995ebf427025cc1f43c00271269addc1ed035c0